Move common/finalizers into util
